refactor: move bill Zod schemas to src/schemas/bill.ts
Separate runtime Zod schemas from pure type definitions by introducing
src/schemas/ for shared validation schemas.
